localdocs: replace hnswlib ANN with usearch exact search
Embeddings are now stored in the SQLite database and a brute-force
search is used to find the best matches.

This fixes situations where the embeddings_v0.dat could get out of sync
with the database.

Also, we now pre-filter the search results instead of post-filtering, so
when you request three snippets you will always get three, even if there
are better matches in unselected collections.

Signed-off-by: Jared Van Bortel <jared@nomic.ai>